requirements.txt
discord
discord.py==1.7.3
numpy<1.23.0,>=1.16.5
requests
pychalk